// src/types.tsx
export interface IAppState  {
    CITY: string
  COUNTRY: string

  TEMP01: string
  ICON01: string
  DATE01: string
  WEATHER01: string
  WIND01: string
  WIND01describe: any

  TEMP02: string
  ICON02: string
  DATE02: string
  WEATHER02: string
  WIND02: string
  WIND02describe: any

  TEMP03: string
  ICON03: string
  DATE03: string
  WEATHER03: string
  WIND03: string
  WIND03describe: any

  TEMP04: string
  ICON04: string
  DATE04: string
  WEATHER04: string
  WIND04: string
  WIND04describe: any
  }